test: share rotated CU session bootstrap
Copilot-Session: 7ccbdc1e-2946-4f0c-b020-b3f1ade57f77
