+++ ddi/tbor/types/tests/harness/mod.rs
@@ -53,8 +53,11 @@ pub mod x509_fixture;
 pub(crate) use azihsm_ddi_tbor_types::build_psk_change_aad;
 pub(crate) use azihsm_ddi_tbor_types::TborPskChangeReq;
 pub(crate) use ctx::TestCtx;
+pub(crate) use fixture::bootstrap_rotated_co;
+pub(crate) use fixture::bootstrap_rotated_cu;
 pub(crate) use fixture::CO_PSK_ID;
 pub(crate) use fixture::CU_PSK_ID;
+pub(crate) use fixture::ROTATED_CO_PSK;
 pub(crate) use fixture::ROTATED_CU_PSK;
 pub(crate) use session::build_mac_fin;
 pub(crate) use session::build_part_init_mach_seed_aad;
